travel-agent-ai/requirements.txt

langchain
langchain-community
langchain-anthropic
anthropic
requests
python-dotenv
pydantic